release 1.2.1

Closes #11, #10, and #12

See merge request luminfire/products/brilliantportal/brilliant-portal-framework!7

+++ routes/api.php
@@ -5,6 +5,7 @@ use BrilliantPortal\Framework\Http\Controllers\Api\Admin\UserController;
 use BrilliantPortal\Framework\Http\Controllers\Api\GenericController;
 use BrilliantPortal\Framework\Http\Middleware\EnsureHasTeam;
 use Illuminate\Support\Facades\Route;
+use Laravel\Jetstream\Features;
 use Vyuldashev\LaravelOpenApi\Generator;
 
 /**
@@ -23,18 +24,21 @@ Route::name('api.')
             ->group(function () {
 
                 Route::apiResource('users', UserController::class);
-                Route::apiResource('teams', TeamController::class)
-                    ->names([
-                        'index' => 'teams.api.index',
-                        'store' => 'teams.api.store',
-                        'destroy' => 'teams.api.destroy',
-                        'update' => 'teams.api.update',
-                        'show' => 'teams.api.show',
-                    ]);
 
-                Route::post('teams/{teamId}/invitations/', [TeamController::class, 'inviteTeamMember'])->name('teams.invitations.create');
-                Route::delete('teams/{teamId}/invitations/{invitationId}', [TeamController::class, 'cancelTeamMemberInvitation'])->name('teams.invitations.cancel');
-                Route::put('teams/{teamId}/remove/{userId}', [TeamController::class, 'removeUser'])->name('teams.invitations.remove');
+                if (Features::hasTeamFeatures()) {
+                    Route::apiResource('teams', TeamController::class)
+                        ->names([
+                            'index' => 'teams.api.index',
+                            'store' => 'teams.api.store',
+                            'destroy' => 'teams.api.destroy',
+                            'update' => 'teams.api.update',
+                            'show' => 'teams.api.show',
+                        ]);
+
+                    Route::post('teams/{teamId}/invitations/', [TeamController::class, 'inviteTeamMember'])->name('teams.invitations.create');
+                    Route::delete('teams/{teamId}/invitations/{invitationId}', [TeamController::class, 'cancelTeamMemberInvitation'])->name('teams.invitations.cancel');
+                    Route::put('teams/{teamId}/remove/{userId}', [TeamController::class, 'removeUser'])->name('teams.invitations.remove');
+                }
             });
 
         /**
